+++ westpa_scripts/robust_runseg_mdrun.py
@@ -18,18 +18,23 @@ walker_str = sys.argv[1].split("/")[-1]
 #backup west.h5
 #note that this is the file at the start of the round (i.e. west-000005.h5 has the data from rounds 1-4)
 #it is important to run this before any segments can complete so that you don't get part of the current round (i.e. the first 23 walkers of round 5)
-os.system(f"cp ../../../west.h5 ../../../h5-backups/west-{round_str}.h5")
+if not os.path.exists(f"../../../h5-backups/west-{round_str}.h5"):
+	os.system(f"cp ../../../west.h5 ../../../h5-backups/west-{round_str}.h5")
+#beware that this does not handle multiple layers of overwriting; it keeps only the most recently replaced file
+else:
+        os.system(f"mv ../../../h5-backups/west-{round_str}.h5 ../../../h5-backups/overwritten-west-{round_str}.h5; cp ../../../west.h5 ../../../h5-backups/west-{round_str}.h5")
 
 
 #lines signalling fatal error requiring restarts
 #found in the log file in seg_logs/
 error_lines = [\
-	"Asynchronous H2D copy failed",\
-	"Asynchronous D2H copy failed",\
-	"The total potential energy is nan, which is not finite.",\
-	"Unexpected cudaStreamQuery failure: an illegal memory access was encountered",\
+ 	"Asynchronous H2D copy failed",\
+ 	"Asynchronous D2H copy failed",\
+ 	"The total potential energy is nan, which is not finite.",\
+ 	"Unexpected cudaStreamQuery failure: an illegal memory access was encountered",\
 	"Could not register the host memory for page locking for GPU transfers.",\
-	"cudaErrorMemoryAllocation: out of memory"]
+	"cudaErrorMemoryAllocation: out of memory",\
+	"bytes failed: out of memory"] #cudaMallocHost of size 4660224 bytes failed: out of memory
 
 #acquire a list of all files present before the simulation is run
 # these are the files not moved to a storage folder when the run has to be restarted
